refactor: RecipeImageUploader 제거 (recipe / review 이미지 저장 경로 동일하게)
