+++ backend/src/main/resources/application-dev.yml
@@ -16,7 +16,3 @@ spring:
 review:
   image:
     path: { DEV_IMAGE_PATH }
-
-recipe:
-  image:
-    path: { DEV_RECIPE_PATH }
