+++ backend/src/test/java/com/funeat/FuneatApplicationTests.java
@@ -9,5 +9,4 @@ class FuneatApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-
 }
