+++ backend/src/test/java/com/funeat/common/TestImageUploader.java
@@ -1,5 +1,6 @@
-package com.funeat.review.application;
+package com.funeat.common;
 
+import com.funeat.review.application.ImageService;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
